% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{Title}
\usage{
anova(waarde, groep)
}
\arguments{
\item{waarde}{The column with measurement values inside a tidy tible}

\item{groep}{The column with the measured groups inside a tidy tibble}
}
\value{
A P-value
}
\description{
Title
}
\examples{
group1 <- "group1"
group2 <- "group2"
group3 <- "group3"
group4 <- "group4"
potato <- read_delim(here::here("data_raw", "02_potato", "potato"),delim=":",locale=locale(decimal_mark=","),col_types=cols(group1=col_number(),group2=col_number(),group3=col_number(),group4=col_number()))
potato_tidy <- pivot_longer(data = read, cols = c(group1, group2, group3, group4),names_to="Groep",values_to="Natrium")
anova(tidy$Natrium, tidy$Groep)
}
