% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem_barchart.R
\name{gem_barchart}
\alias{gem_barchart}
\title{Title}
\usage{
gem_barchart(data, groep, waarde)
}
\arguments{
\item{data}{A tidy tibble with group and values column}

\item{groep}{The column with the measured groups}

\item{waarde}{The column with measurement values}
}
\value{
A barchart with for each param groep the mean param kolom with error bars of standard deviation 1.
}
\description{
Title
}
\examples{
rabbit_ears <- data.frame(rabbitnr = seq(1,20),colour = c(rep("beige",10),rep("brown",10)), earlength = c(17.5,13.0,18.6,19.4,20.9,18.9,22.8,20.3,22.1,20.4,23.0,24.1,26.1,25.2,27.8,28.5,26.5,21.5,20.2,18.8))
gem_barchart(rabbit_ears, colour, earlength)
}
