% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantGrowth_edit.R
\docType{data}
\name{PlantGrowth_edit}
\alias{PlantGrowth_edit}
\title{Plant growth data edited with height and leafs values}
\format{
An Rds table file
}
\source{
...
}
\usage{
data(PlantGrowth_edit)
}
\description{
Dummy data generated by my DSFB2 teachers, to which I added height and leaf values so this data can be used as an universal function example.
}
\examples{
data(PlantGrowth_edit)
library("ggplot2")
cor_scatterplot(PlantGrowth_edit, PlantGrowth_edit$leafs, PlantGrowth_edit$height, reposition_r = 4)
}
\references{
...
}
\keyword{datasets}
