% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato.R
\docType{data}
\name{potato}
\alias{potato}
\title{Potato data}
\format{
An Rds table file
}
\source{
...
}
\usage{
data(potato)
}
\description{
Dummy data generated by my DSFB2 teachers. It contains the sodium contents for four groups of chips.
}
\examples{
data(potato)
anova(potato$Natrium, potato$Groep)
}
\references{
...
}
\keyword{datasets}
