% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_scatterplot.R
\name{cor_scatterplot}
\alias{cor_scatterplot}
\title{Title}
\usage{
cor_scatterplot(data, kolom_x, kolom_y, groep)
}
\arguments{
\item{data}{A tidy tibble with grouping and two value columns.}

\item{kolom_x}{The measurement values of a variable which are to correlate to...}

\item{kolom_y}{...the measurement values of another variable}

\item{groep}{The groups into which we can divide the values}
}
\value{
A scatterplot to visualize the correlation between twe two variables, with a Pearson's correlation coefficient
}
\description{
Title
}
\examples{
library("palmerpenguins")
cor_scatterplot(penguins, penguins$body_mass_g, penguins$flipper_length_mm, penguins$species)
}
